work in progress for sharing
