+++ src/introspector/input.py
@@ -1,0 +1,4 @@
+def inputs(dataset):
+    for x in ("inputa","inputb"):
+        yield dataset + x
+#app_id = None
