+++ src/introspector/orient.py
@@ -1,0 +1,5 @@
+def orient(data):
+    toemoji(data)
+    yield from summarize(
+        sort(
+            filtering(data)))  # show a summary of the data
