+++ src/introspector/page_size.py
@@ -1,0 +1,1 @@
+from app_args import get_page_size
