+++ src/introspector/summarize.py
@@ -1,0 +1,17 @@
+
+def summarize(data):
+
+    # lets see if we can use emojis to summarize.
+    toemoji(data)
+    
+    #if isinstance(data, generato):
+    if isinstance(data, Iterable):
+        if isinstance(data, types.GeneratorType):
+            pass
+        else:
+            st.write("Sum Object is iterable", type(data).__name__, data, )
+        for x in data:
+            yield x
+    else:
+        st.write("Sum Object not an iterable")
+        yield data
